+++ frontend/app/chat/page.tsx
@@ -5,12 +5,9 @@ import { Button } from "@/components/ui/button";
 import { useEffect, useRef, useState } from "react";
 import { Separator } from "@/components/ui/separator";
 import { users, messages } from "./test-data";
-import type { Message, User } from "./test-data";
-import {
-  AvatarSkeleton,
-  SmallAvatarSkeleton,
-  MessageSkeleton,
-} from "./skeleton";
+import type { Message } from "./test-data";
+import { AvatarSkeleton, MessageSkeleton } from "./skeleton";
+import { Sidebar } from "./sidebar";
 
 function MessageItem({
   message,
@@ -60,33 +57,6 @@ function MessageGroup({ messages }: { messages: Message[] }) {
         return <MessageItem message={msg} withAvatar={i === 0} key={msg.id} />;
       })}
     </Stack>
-  );
-}
-
-function truncateString(str: string, num: number): string {
-  if (str.length <= num) {
-    return str;
-  }
-  return str.slice(0, num) + "...";
-}
-
-function Sidebar({ users }: { users: User[] }) {
-  return (
-    <div className="overflow-y-auto shrink-0 basis-36 flex-grow pb-4">
-      <Stack spacing={2}>
-        {users.map((user) => (
-          <button
-            key={user.id}
-            className="flex gap-2 items-center group hover:opacity-60"
-          >
-            <SmallAvatarSkeleton />
-            <a className="text-muted-foreground text-sm whitespace-nowrap group-hover:text-primary">
-              {truncateString(user.name, 15)}
-            </a>
-          </button>
-        ))}
-      </Stack>
-    </div>
   );
 }
